Drafting plant and seed interactions
