+++ Assets/Scripts/CharacterController/Interactions/Seeds.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cb12cf7bf01c59e4b931e2eee869b9be
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
