+++ Assets/Scripts/CharacterController/Interactions/Seeds/PlantInteractions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 204ad4e41b2dbff46ba76776517897a5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
